requirements.txt
cffi==1.15.1
Pillow==9.5.0
pillow-heif==0.11.0
pycparser==2.21
moviepy
ffmpeg-python